Reject requests with unsupported methods
Currently, such methods lead to a match error in the routeTrie
lookup. This exception is printed, but since undertow defaults to
responding with 200, the effect is that requests with invalid methods
always return a successful response code.

The changes here modify this behavior to respond with a 405 if the
dispatch trie cannot be found in the method map.