Allow overlap between static routes and wildcards
This changes the DispatchTrie to allow overlapping wildcard segments
in paths with static ones, with a preference for the latter.

For example, consider the following routes:

```
@cask.get("/settings")
def settings() = "settings"

@cask.get("/:user")
def user(id: String) = id
```

This is currently not allowed. With these changes, it would be
allowed, and the static route `settings` would be preffered, with a
fallback to the dynamic route `user`:

```
GET /settings => settings
GET /foo => foo
GET /bar => bar
```

---

The reason I'm proposing this change is mostly for use in HTML
applications (i.e. not RPC-style JSON APIs). In this scenario, short
URLs are useful, since users may type them directly and associate
meaning to them.

Consider for example the way GitHub structures URLs. If github were
written with cask's current routing logic, it would not be possible to
have URLs such as `/settings` and `/com-lihaoyi`, and instead some
namespacing would need to be introduced (e.g. /orgs/com-lihaoyi/) to
separate these, which might not actually be relevant for users.